% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunctions.R
\name{binint}
\alias{binint}
\title{Auxiliar function to transform a number to a binary vector}
\usage{
binint(x, D = ceiling(log(x, 2)))
}
\arguments{
\item{x}{Integer to transform}

\item{D}{Significant digits}
}
\value{
Binary vector with binary representation of the integer
}
\description{
Auxiliar function to transform a number to a binary vector
}
