% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montecarlo.R
\name{fsearch}
\alias{fsearch}
\title{Full Blind Search Method}
\usage{
fsearch(search, FUN, type = "min", ...)
}
\arguments{
\item{search}{Matrix with solutions x D}

\item{FUN}{Evaluation function}

\item{type}{Type of the searched optimum}

\item{...}{Extra arguments for FUN function}
}
\value{
List
   \item{index}{Index of element with min/max result}
   \item{sol}{Optimum location}
   \item{eval}{Optimum}
}
\description{
\code{fsearch} evaluates a function in the given points and returns the optimum
}
