% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill.R
\name{hclimbing}
\alias{hclimbing}
\title{Hill Climbing Method}
\usage{
hclimbing(par, fun, change, lower, upper, control, type = "min", ...)
}
\arguments{
\item{par}{Initial solution}

\item{fun}{Evaluation function}

\item{change}{Function to generate the next candidate}

\item{lower}{Vector with lowest values for each dimension}

\item{upper}{Vector with highest values for each dimension}

\item{control}{List with stopping and monitoring method: maxit (maximum number of iterations), REPORT (frequency of monitoring information)}

\item{type}{Optimization type "min" or "max"}

\item{...}{Extra parameters for FUN}
}
\value{
List with solution and evaluation
}
\description{
Hill Climbing Method
}
