% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill.R
\name{hclimbingTSP}
\alias{hclimbingTSP}
\title{Hill Climbing Method}
\usage{
hclimbingTSP(par, change, coordinates, control, ...)
}
\arguments{
\item{par}{Initial solution}

\item{change}{Function to generate the next candidate}

\item{coordinates}{Coordinates}

\item{control}{List with stopping and monitoring method: maxit (maximum number of iterations), REPORT (frequency of monitoring information)}
}
\value{
List with solution and evaluation
}
\description{
Hill Climbing Method
}
