% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunctions.R
\name{intbin}
\alias{intbin}
\title{Auxiliar function to transform a binary vector into its decimal representation}
\usage{
intbin(x)
}
\arguments{
\item{x}{Binary vector}
}
\value{
Decimal representation
}
\description{
Auxiliar function to transform a binary vector into its decimal representation
}
