% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montecarlo.R
\name{mcsearch}
\alias{mcsearch}
\title{Montecarlo Uniform Search Method}
\usage{
mcsearch(N, lower, upper, FUN, type = "min", ...)
}
\arguments{
\item{N}{Number of samples}

\item{lower}{Vector with lowest values for each dimension}

\item{upper}{Vector with highest values for each dimension}

\item{FUN}{Evaluation function}

\item{type}{Type of the searched optimum}

\item{...}{Extra arguments for FUN function}
}
\value{
List
   \item{index}{Index of element with min/max result}
   \item{sol}{Optimum location}
   \item{eval}{Optimum}
}
\description{
\code{mcsearch} evaluates a function using Montecarlo method and returns the optimum
}
