% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montecarlo.R
\name{mcsearchTSP}
\alias{mcsearchTSP}
\title{Montecarlo Uniform Search Method for TSP}
\usage{
mcsearchTSP(N, coordinates, ...)
}
\arguments{
\item{N}{Number of samples}

\item{coordinates}{TSP coordinates}

\item{...}{Extra arguments for FUN function}
}
\value{
List
   \item{index}{Index of element with min result}
   \item{sol}{Optimum location}
   \item{eval}{Optimum}
}
\description{
\code{mcsearchTSP} evaluates a function using Montecarlo method and returns the optimum
}
