% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSP.R
\name{plotTSP}
\alias{plotTSP}
\title{Travel Salesperson Problem Function}
\usage{
plotTSP(x, coordinates)
}
\arguments{
\item{x}{Proposed solution}

\item{coordinates}{Matrix with coordinates}
}
\value{
Length of the travel
}
\description{
Travel Salesperson Problem Function
}
