% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunctions.R
\name{sumbin}
\alias{sumbin}
\title{sumbin of binary raw object x}
\usage{
sumbin(x)
}
\arguments{
\item{x}{Binary vector}
}
\value{
sum of the bits of the vector
}
\description{
sumbin of binary raw object x
}
