% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSP.R
\name{tsp.evaluation}
\alias{tsp.evaluation}
\title{Travel Salesperson Problem Function}
\usage{
tsp.evaluation(x, coordinates)
}
\arguments{
\item{x}{Proposed solution}

\item{coordinates}{Matrix with coordinates}
}
\value{
Length of the travel
}
\description{
Travel Salesperson Problem Function
}
